Move sample views to according package
