+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.PokemonAndroidOldTech"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".ui.views.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.PokemonAndroidOldTech">
             <intent-filter>
